requirements.txt
aiogram==3.*
aiohttp
python-dotenv
supabase
